feat: replace mock vault data with live yield-opportunities API
