+++ packages/contracts/contracts/vault/src/test.rs.rej
@@ -1,0 +1,31 @@
+@@ -1438,6 +1438,30 @@
+ }
+ 
+ #[test]
++fn test_harvest_impairment_zero_fee() {
++    let (env, admin, token, vault, treasury) = setup();
++    let user = Address::generate(&env);
++    let deposit = 1_000 * XLM;
++    mint(&token, &user, deposit);
++    vault.deposit(&user, &deposit, &0);
++
++    vault.grant_role(&admin, &admin, &Role::Manager);
++
++    // Give them some pending yield first
++    vault.report_yield(&admin, &(100 * XLM));
++    
++    // Simulate impairment by changing total assets directly so redeemable < principal
++    // wait, report_yield with a huge negative number reduces total assets.
++    vault.report_yield(&admin, &(-(300 * XLM)));
++    
++    // Harvest
++    let result = vault.harvest(&user);
++    assert_eq!(result.gross_yield, 0);
++    assert_eq!(result.performance_fee, 0);
++    assert_eq!(result.net_yield, 0);
++}
++
++#[test]
+ fn test_harvest_impairment_no_fee_charged() {
+     // When yield is negative (impairment), no performance fee should be charged
+
